Layout modifications and Uservoice add

+++ Joobsbox/Model/Search.php
@@ -21,29 +21,39 @@
 
 class Joobsbox_Model_Search {
 	public $_index;
+	public $_enabled = true;
 	
 	public function __construct() {
 		if(file_exists("Joobsbox/SearchIndexes/main")) {
 			$this->_index = Zend_Search_Lucene::open("Joobsbox/SearchIndexes/main");
 		} else {
-			$this->_index = Zend_Search_Lucene::create("Joobsbox/SearchIndexes/main");
+		  if(is_writable('Joobsbox/SearchIndexes')) {
+			  $this->_index = Zend_Search_Lucene::create("Joobsbox/SearchIndexes/main");
+			} else {
+			  $this->_enabled = false;
+			}
 		}
-		Zend_Search_Lucene_Analysis_Analyzer::setDefault(new Zend_Search_Lucene_Analysis_Analyzer_Common_Utf8_CaseInsensitive ()); 
-		Zend_Search_Lucene_Search_QueryParser::setDefaultEncoding('utf-8'); 
+		if($this->_enabled) {
+		  Zend_Search_Lucene_Analysis_Analyzer::setDefault(new Zend_Search_Lucene_Analysis_Analyzer_Common_Utf8_CaseInsensitive ()); 
+		  Zend_Search_Lucene_Search_QueryParser::setDefaultEncoding('utf-8'); 
+		}
 	}
 	
 	public function search($string) {
+	  if(!$this->_enabled) return array();
 		$query = Zend_Search_Lucene_Search_QueryParser::parse($string);
 		return $this->_index->find($query);
 	}
 	
 	public function searchTag($tag, $value) {
+	  if(!$this->_enabled) return array();
 	  Zend_Search_Lucene::setDefaultSearchField($tag);
 	  $query = Zend_Search_Lucene_Search_QueryParser::parse($value);
 		return $this->_index->find($query);
 	}
 	
 	public function deleteJob($jobId) {
+	  if(!$this->_enabled) return false();
 		$term = new Zend_Search_Lucene_Index_Term($jobId, 'ID');
 		$hits  = $this->_index->termDocs($term);
 		if(count($hits)) {
@@ -54,6 +64,7 @@ class Joobsbox_Model_Search {
 	}
 	
 	public function addJob($jobData) {
+	  if(!$this->_enabled) return false();
 		// Delete old job with the same id from index
 		$term = new Zend_Search_Lucene_Index_Term($jobData['ID'], 'ID');
 		$hits  = $this->_index->termDocs($term);
@@ -78,6 +89,7 @@ class Joobsbox_Model_Search {
 	}
 	
 	public function resetIndex() {
+	  if(!$this->_enabled) return false();
 		for ($count = 0; $count < $this->_index->count(); $count++) {
         $this->_index->delete($count);
     }
@@ -85,6 +97,7 @@ class Joobsbox_Model_Search {
 	}
 	
 	public function commit() {
+	  if(!$this->_enabled) return false();
 		$this->_index->commit();
 		$this->_index->optimize();
 	}
